Use SharingPool to share Django postgres connections with sqlalchemy.
